ERA Concept Hierarchy Tables

Creates a list of data.tables describing the organizational structure of the ERA concept hierarchy for use in other ERA functions and interactive apps.

CreateConceptLevels()

Value

CreateConceptLevels returns a list of four elements each a data.table:

Prac.Levels Practice concept organizational levels

Out.Levels Outcome concept organizational levels

Prod.Levels Product (also called EU or experimental units) concept organizational levels

Agg.Levels This data.table describes fields that can be used to aggregate data. Observation = no aggregation; each row of ERA is an observation, Study = the publication/study an observation is reported from, and Location the spatial location of an observation (a combination of latitude, longitude and a buffer of uncertainty). Choice and Choice.Code fields are simplified names/codes for hierarchical concept levels for user selection task. The remaining columns indicate the corresponding column names in the ERA dataset for user selections.