The compiled ERA dataset (version = Comb 2023.04.26 2023.06.27)

Index

unique row identity

Code

unique publication identity

Author

primary author of the publication

Date

the year of publication

Journal

abbreviation of the journal name

DOI

doi (or url if no doi available) of article. DOI = digital object identifier

Elevation

the elevation in meters (m). Provide the mid-point if a range is given

Country

country name

ISO.3166-1.alpha-3

Standardized country name using ISO.3166-1.alpha-3. This field should be auto filled using the country selected and data in the Levels Tab

Site.Type

one of the following: Farm, Station, Greenhouse, Survey, or Lab. Farm research is conducted in a farmers field and can be managed by the farmer or researcher. Station research is conducted in a controlled setting of research station, university or school. Survey research is conducted via interviews that yields quantitative data based on testimonial. Greenhouse is conducted in a greenhouse and is only relevant for studies of greenhouse gas emissions data. Lab studies can include fisheries feeding trials conducted at a small scale in university research labs

Site.ID

free text name to identify the site. Include institution/station name and location or village/town in county/district, but not the country

MAT

reported mean annual temperature in degrees C. If a range is given, average the values and enter the average. If mean highs and lows are given, they can be averaged to estimate the mean annual temperature

MAP

mean annual precipitation (mm) for the location. This should be a long-term average for more than 10 years of data

TAP

total annual precipitation (mm) for the location in the year corresponding to the outcome data

MSP

mean seasonal precipitation (mm) for the location in the season corresponding to the outcome data. This should be a long-term average from more than 10 years of data

TSP

total seasonal precipitation (mm) for the location in the season corresponding to the outcome data

Soil.Type

list the given soil classification/name in lowercase letters. If more than one soil type is given (within the same taxonomic system) enter both separated by a period

Soil.Classification

list the taxonomic system used to classify the soil. Choose from (FAO, USDA, other)

Soil.Texture

list the soil texture provided in the text or in tables in lowercase letters. If no texture is named, but % of sand, silt and clay are given, use the soil texture triangle to estimate soil texture (https://www.nrcs.usda.gov/wps/portal/nrcs/detail/soils/survey/?cid=nrcs142p2_054167)

SOC

soil organic carbon value derived during site characterization

SOC.Unit

units provided for soil organic carbon. Usually either g/kg or % (also can be density kg/m3 or stocks t/ha)

SOC.Depth

concatenate min (top) and max (bottom) depths user entered with a "-", e.g. 0-10

Soil.pH

soil pH value derived during site characterization

Soil.pH.Method

method used to calculate soil pH

Plant.Start

starting of planting period (dd.mm.yyyy)

Plant.End

end of planting period (dd.mm.yyyy)

Harvest.Start

starting of harvest period (dd.mm.yyyy)

Harvest.End

end of harvest period (dd.mm.yyyy)

Rep

the number of replicates for the treatments

Plot.Size

the size of the smallest treatment plot unit in m2

CID

a unique identifier (within study) describing a set of management practices that occurred in the control which is used to identify the treatment combinations tested, to allow comparison across outcomes in a paper. A default system of T1, T2, T3, etc. can be used. CID should describe the treatment used as the control for this observation

C.Descrip

(superceded by C.Descrip.Clean column) Free text description of the control treatment

C.NI

amount of inorganic (chemical) nitrogen applied to control plots in kg/ha

C.NO

amount of nitrogen in organic materials (manures or compost, but not residues) applied to control plots, preferably in kg/ha

TID

a unique identifier (within study) describing a set of management practices that occurred in treatment which is used to identify the treatment combinations tested, to allow comparison across outcomes in a paper. A default system of T1, T2, T3, etc. can be used. TID should describe the treatment used as the control for this observation

T.Descrip

(superceded by T.Descrip.Clean column) Free text description of the CSA treatment

T.NI

amount of inorganic (chemical) nitrogen (N) applied to CSA treatment in kg/ha

T.NO

amount of nitrogen (N) in organic materials (manures or compost) applied to CSA treatment, preferably in kg/ha

Diversity

name of species in intercrop or rotations including the crop. Dash(-) indicates intercrop and slash (/) indicates a rotation

Variety

commercial name of the crop or animal variety

Tree

Name of tree species used in agroforestry experiments, can be both as an intercrop or as an agroforestry mulch

Duration

length of time the experiment has been ongoing up to the date the data are being reported for in years. In bimodal rainfall systems with multiple growing seasons, seasons are identified by fractions of the year, 0.5, etc. If outcomes are reported in a unimodal rainfall system with one growing season per year, the duration can be equivalent to one year

M.Year

the calendar year the measurement was taken. Seasons in multi-season systems are identified by decimals .1 or .2 for the first and second season respectively. When the measurements are average values for more than one year create a decimal for the years covered

M.Year.Start

first year the measurement was taken (YYYY). For yield or similar outcomes this should relate to the harvest year (rather than the planting year)

M.Year.End

this only required for temporally aggregated observations (e.g. a result that has been averaged from crop yields reported from 2000,2001 and 2002)

Season.Start

this only required in areas where there are mutiple growing seasons in a year (e.g., areas with bimodal rainfall or irrigated cropping in dry seasons of a unimodal system)

Season.End

this only required for temporally aggregated observations in areas with more than one growing season (e.g. a result that has been averaged from crop yields reported from 2000 season 1,2001 season 1 and 2002 season 1)

EU

the experimental unit (EU) for which experiment outcomes are reported. This usually a crop or animal product. For cookstove papers it may be a cookstove. The EU for soil outcomes should be the crop(s) grown on the soil unless no crops are grown at all. Letters in the code should be lowercase. If the outcome refers to multiple experimental units (e.g. soil outcomes during an intercropping experiment) select all experimental units separated by a period (see EU2 tab)

Outcode

the productivity, resilience, or mitigation outcome indicator that data are being reported on (see OUTCOMES2 tab)

Units

free text description of the units of the data being reported without special characters or formatting

MeanC

the value that corresponds to the control treatment and outcome. Please standard yields to t/ha where appropropriate, keeping as many decimal places as there is data for

MeanT

the value that corresponds to the CSA treatment and outcome. Please standard yields to kg/ha where appropropriate, keeping as many decimal places as there is data for

Upper

the depth for which the MeanC and MeanT correspond for soil properties in cm (lowest depth value = upper, so if you have a depth of 0-20 cm the lowest value = 20)

Lower

the depth for which the MeanC and MeanT correspond for soil properties in cm (highest depth value = lower, so if you have a depth of 0-20 cm the highest value = 0)

DataLoc

the location in the paper where the data can be found. This should be three letters (fig, tab, or txt) in lowercase and a number (for figures and tables) to signify figure, table or text, respectively

USD2010.C

monetary outcomes converted to international USD in 2010 using World Bank CPI, PPP and exchange rates. There are no data for Zimbabwe.

USD2010.T

monetary outcomes converted to international USD in 2010 using World Bank CPI, PPP and exchange rates. There are no data for Zimbabwe.

MeanFlip

is a negative result considered better than a positive result (i.e. MeanT/MeanC <1 = good)? This is indexed in using the Outcode field in the compednium and the Negative Values field in the OUTCOMES tab

T.Feed.Source

where were livestock diets in the treatment practice(s) sourced from?

C.Feed.Source

Where were livestock diets in the control practice(s) sourced from?

Species

the tree species for which the experiment outcome are reported

Partial.Outcome.Name

free text to describe the partial outcome

Partial.Outcome.Code

ERA practice codes that correspond to the partial outcome

Analysis.Function

Describes the logical process used to compare control to treatment (for developers)

Version

year in which the ERA data entry took place. Either 2018 or 2020

Mulch.Code

ERA mulch code based on the nitrogen fixing properties specified in Nfix. NA for animal products

Irrigation.C

indicates the presence of irrigation in the control treatment

Irrigation.T

indicates the presence of irrigation in the experimental treatment

Irrigation.Meth.C

method of irrigation used in the control treatment. Cell will be blank if Irrigation.C is FALSE

Irrigation.Meth.T

method of irrigation used in the experimental treatment. Cell will be blank if Irrigation.T is FALSE

C1:C13

codes corresponding to all practices used on the control treatment described in C.Descrip. Only one code is placed in each column (see subpractices column of PRACTICES2 tab)

T1:T13

codes corresponding to all practices used on the CSA treatment described in T.Descrip. Only one code is placed in each column (see subpractices column of PRACTICES2 tab)

ID

study number

obs_count

number of observations

out_count

number of outcomes reported per study

Lat

latitude of study location in degrees or decimal degrees, the latter recorded to same accuracy as in publication (or estimated from googlemaps where locations were found from site names/descriptions). For decimal degrees the sign is not recorded here, but in the LatH column instead (all values should be positive)

Lon

longitude of study location in decimal degrees, recorded to same accuracy as in publication (or estimated from googlemaps where locations were found from site names or descriptions). For decimal degrees the sign is not recorded here, but in the LatH column instead. All values should be positive.

yi

the natural log of the response ratio log(MeanT/MeanC)

pc

pc=proportional change.It is a measure of the ratio of experimental to control outcomes

plist

experimental ERA practices that are in the experimental, but not the control treatment

base.list

base practices that are shared between experimental and control treatments

EUlist

the product code OutcomeCodes$Code for the observation using a - delim for multiple products

Out.Pillar

outcome pillar name

Out.SubPillar

outcome subpillar name

Out.Ind

outcome indicator name

Out.SubInd

outcome subindicator name

Out.SubInd

outcome subindicator name simplified

SubPrName

experimental subpractice name (note that this corresponds to the Subpractice.S field in the PracticeCodes)

PrName

experimental practice name

Theme

experimental practice theme name

SubPrName.Base

base practice name (note that this corresponds to the Subpractice.S field in the PracticeCodes)

PrName.Base

base practice name

Theme.Base

base practice theme name

Product

the full name of the product, including aspect of the component such as grain, meat or milk

Product.Type

product type name

Product.Subtype

product subtype name

Product.Simple

simplified product name, excluding any aspects of component

Latitude

latitude of study location in degrees or decimal degrees

Longitude

longitude of study location in degrees or decimal degrees

Buffer

manual estimation (e.g. from measuring on google earth) in m of the spatial uncertainty for a point location. The site should within an X m radius (the buffer.manual amount) of the point location specified. Unit is meters

Site.Key

longitude, latitude and buffer for the site

AEZ16simple

simplified version of AEZ16 removing the tropic/subtropic classification

AEZ16

AEZs for Africa South of the Sahara (SSA) based on the methodology developed by FAO and IIASA, using the scheme with 16 classes

AEZ5

AEZs Africa South of the Sahara (SSA) based on the methodology developed by FAO and IIASA, using the scheme with 5 classes

Mean.Annual.Precip

mean annual precipitation (mm) derived from the CHIRPS 2.0 dataset API at 0.05&deg resolution

Mean.Annual.Temp

mean annual temperature (C) derived from the NASA POWER dataset API at 0.5&deg resolution

CLY

weight percentage of clay particles (<0.0002 mm)

SLT

weight percentage of silt particles (0.0002-0.05 mm)

SND

weight percentage of the sand particles (0.05-2 mm)

...

The compiled ERA dataset (version = Comb 2023.04.26 2023.06.27)

Index

unique row identity

Code

unique publication identity

Author

primary author of the publication

Date

the year of publication

Journal

abbreviation of the journal name

DOI

doi (or url if no doi available) of article. DOI = digital object identifier

Elevation

the elevation in meters (m). Provide the mid-point if a range is given

Country

country name

ISO.3166-1.alpha-3

Standardized country name using ISO.3166-1.alpha-3. This field should be auto filled using the country selected and data in the Levels Tab

Site.Type

one of the following: Farm, Station, Greenhouse, Survey, or Lab. Farm research is conducted in a farmers field and can be managed by the farmer or researcher. Station research is conducted in a controlled setting of research station, university or school. Survey research is conducted via interviews that yields quantitative data based on testimonial. Greenhouse is conducted in a greenhouse and is only relevant for studies of greenhouse gas emissions data. Lab studies can include fisheries feeding trials conducted at a small scale in university research labs

Site.ID

free text name to identify the site. Include institution/station name and location or village/town in county/district, but not the country

MAT

reported mean annual temperature in degrees C. If a range is given, average the values and enter the average. If mean highs and lows are given, they can be averaged to estimate the mean annual temperature

MAP

mean annual precipitation (mm) for the location. This should be a long-term average for more than 10 years of data

TAP

total annual precipitation (mm) for the location in the year corresponding to the outcome data

MSP

mean seasonal precipitation (mm) for the location in the season corresponding to the outcome data. This should be a long-term average from more than 10 years of data

TSP

total seasonal precipitation (mm) for the location in the season corresponding to the outcome data

Soil.Type

list the given soil classification/name in lowercase letters. If more than one soil type is given (within the same taxonomic system) enter both separated by a period

Soil.Classification

list the taxonomic system used to classify the soil. Choose from (FAO, USDA, other)

Soil.Texture

list the soil texture provided in the text or in tables in lowercase letters. If no texture is named, but % of sand, silt and clay are given, use the soil texture triangle to estimate soil texture (https://www.nrcs.usda.gov/wps/portal/nrcs/detail/soils/survey/?cid=nrcs142p2_054167)

SOC

soil organic carbon value derived during site characterization

SOC.Unit

units provided for soil organic carbon. Usually either g/kg or % (also can be density kg/m3 or stocks t/ha)

SOC.Depth

concatenate min (top) and max (bottom) depths user entered with a "-", e.g. 0-10

Soil.pH

soil pH value derived during site characterization

Soil.pH.Method

method used to calculate soil pH

Plant.Start

starting of planting period (dd.mm.yyyy)

Plant.End

end of planting period (dd.mm.yyyy)

Harvest.Start

starting of harvest period (dd.mm.yyyy)

Harvest.End

end of harvest period (dd.mm.yyyy)

Rep

the number of replicates for the treatments

Plot.Size

the size of the smallest treatment plot unit in m2

CID

a unique identifier (within study) describing a set of management practices that occurred in the control which is used to identify the treatment combinations tested, to allow comparison across outcomes in a paper. A default system of T1, T2, T3, etc. can be used. CID should describe the treatment used as the control for this observation

C.Descrip

(superceded by C.Descrip.Clean column) Free text description of the control treatment

C.NI

amount of inorganic (chemical) nitrogen applied to control plots in kg/ha

C.NO

amount of nitrogen in organic materials (manures or compost, but not residues) applied to control plots, preferably in kg/ha

TID

a unique identifier (within study) describing a set of management practices that occurred in treatment which is used to identify the treatment combinations tested, to allow comparison across outcomes in a paper. A default system of T1, T2, T3, etc. can be used. TID should describe the treatment used as the control for this observation

T.Descrip

(superceded by T.Descrip.Clean column) Free text description of the CSA treatment

T.NI

amount of inorganic (chemical) nitrogen (N) applied to CSA treatment in kg/ha

T.NO

amount of nitrogen (N) in organic materials (manures or compost) applied to CSA treatment, preferably in kg/ha

Diversity

name of species in intercrop or rotations including the crop. Dash(-) indicates intercrop and slash (/) indicates a rotation

Variety

commercial name of the crop or animal variety

Tree

Name of tree species used in agroforestry experiments, can be both as an intercrop or as an agroforestry mulch

Duration

length of time the experiment has been ongoing up to the date the data are being reported for in years. In bimodal rainfall systems with multiple growing seasons, seasons are identified by fractions of the year, 0.5, etc. If outcomes are reported in a unimodal rainfall system with one growing season per year, the duration can be equivalent to one year

M.Year

the calendar year the measurement was taken. Seasons in multi-season systems are identified by decimals .1 or .2 for the first and second season respectively. When the measurements are average values for more than one year create a decimal for the years covered

M.Year.Start

first year the measurement was taken (YYYY). For yield or similar outcomes this should relate to the harvest year (rather than the planting year)

M.Year.End

this only required for temporally aggregated observations (e.g. a result that has been averaged from crop yields reported from 2000,2001 and 2002)

Season.Start

this only required in areas where there are mutiple growing seasons in a year (e.g., areas with bimodal rainfall or irrigated cropping in dry seasons of a unimodal system)

Season.End

this only required for temporally aggregated observations in areas with more than one growing season (e.g. a result that has been averaged from crop yields reported from 2000 season 1,2001 season 1 and 2002 season 1)

EU

the experimental unit (EU) for which experiment outcomes are reported. This usually a crop or animal product. For cookstove papers it may be a cookstove. The EU for soil outcomes should be the crop(s) grown on the soil unless no crops are grown at all. Letters in the code should be lowercase. If the outcome refers to multiple experimental units (e.g. soil outcomes during an intercropping experiment) select all experimental units separated by a period (see EU2 tab)

Outcode

the productivity, resilience, or mitigation outcome indicator that data are being reported on (see OUTCOMES2 tab)

Units

free text description of the units of the data being reported without special characters or formatting

MeanC

the value that corresponds to the control treatment and outcome. Please standard yields to t/ha where appropropriate, keeping as many decimal places as there is data for

MeanT

the value that corresponds to the CSA treatment and outcome. Please standard yields to kg/ha where appropropriate, keeping as many decimal places as there is data for

Upper

the depth for which the MeanC and MeanT correspond for soil properties in cm (lowest depth value = upper, so if you have a depth of 0-20 cm the lowest value = 20)

Lower

the depth for which the MeanC and MeanT correspond for soil properties in cm (highest depth value = lower, so if you have a depth of 0-20 cm the highest value = 0)

DataLoc

the location in the paper where the data can be found. This should be three letters (fig, tab, or txt) in lowercase and a number (for figures and tables) to signify figure, table or text, respectively

USD2010.C

monetary outcomes converted to international USD in 2010 using World Bank CPI, PPP and exchange rates. There are no data for Zimbabwe.

USD2010.T

monetary outcomes converted to international USD in 2010 using World Bank CPI, PPP and exchange rates. There are no data for Zimbabwe.

MeanFlip

is a negative result considered better than a positive result (i.e. MeanT/MeanC <1 = good)? This is indexed in using the Outcode field in the compednium and the Negative Values field in the OUTCOMES tab

T.Feed.Source

where were livestock diets in the treatment practice(s) sourced from?

C.Feed.Source

Where were livestock diets in the control practice(s) sourced from?

Species

the tree species for which the experiment outcome are reported

Partial.Outcome.Name

free text to describe the partial outcome

Partial.Outcome.Code

ERA practice codes that correspond to the partial outcome

Analysis.Function

Describes the logical process used to compare control to treatment (for developers)

Version

year in which the ERA data entry took place. Either 2018 or 2020

Mulch.Code

ERA mulch code based on the nitrogen fixing properties specified in Nfix. NA for animal products

Irrigation.C

indicates the presence of irrigation in the control treatment

Irrigation.T

indicates the presence of irrigation in the experimental treatment

Irrigation.Meth.C

method of irrigation used in the control treatment. Cell will be blank if Irrigation.C is FALSE

Irrigation.Meth.T

method of irrigation used in the experimental treatment. Cell will be blank if Irrigation.T is FALSE

C1:C13

codes corresponding to all practices used on the control treatment described in C.Descrip. Only one code is placed in each column (see subpractices column of PRACTICES2 tab)

T1:T13

codes corresponding to all practices used on the CSA treatment described in T.Descrip. Only one code is placed in each column (see subpractices column of PRACTICES2 tab)

ID

study number

obs_count

number of observations

out_count

number of outcomes reported per study

Lat

latitude of study location in degrees or decimal degrees, the latter recorded to same accuracy as in publication (or estimated from googlemaps where locations were found from site names/descriptions). For decimal degrees the sign is not recorded here, but in the LatH column instead (all values should be positive)

Lon

longitude of study location in decimal degrees, recorded to same accuracy as in publication (or estimated from googlemaps where locations were found from site names or descriptions). For decimal degrees the sign is not recorded here, but in the LatH column instead. All values should be positive.

yi

the natural log of the response ratio log(MeanT/MeanC)

pc

pc=proportional change.It is a measure of the ratio of experimental to control outcomes

plist

experimental ERA practices that are in the experimental, but not the control treatment

base.list

base practices that are shared between experimental and control treatments

EUlist

the product code OutcomeCodes$Code for the observation using a - delim for multiple products

Out.Pillar

outcome pillar name

Out.SubPillar

outcome subpillar name

Out.Ind

outcome indicator name

Out.SubInd

outcome subindicator name

Out.SubInd

outcome subindicator name simplified

SubPrName

experimental subpractice name (note that this corresponds to the Subpractice.S field in the PracticeCodes)

PrName

experimental practice name

Theme

experimental practice theme name

SubPrName.Base

base practice name (note that this corresponds to the Subpractice.S field in the PracticeCodes)

PrName.Base

base practice name

Theme.Base

base practice theme name

Product

the full name of the product, including aspect of the component such as grain, meat or milk

Product.Type

product type name

Product.Subtype

product subtype name

Product.Simple

simplified product name, excluding any aspects of component

Latitude

latitude of study location in degrees or decimal degrees

Longitude

longitude of study location in degrees or decimal degrees

Buffer

manual estimation (e.g. from measuring on google earth) in m of the spatial uncertainty for a point location. The site should within an X m radius (the buffer.manual amount) of the point location specified. Unit is meters

Site.Key

longitude, latitude and buffer for the site

AEZ16simple

simplified version of AEZ16 removing the tropic/subtropic classification

AEZ16

AEZs for Africa South of the Sahara (SSA) based on the methodology developed by FAO and IIASA, using the scheme with 16 classes

AEZ5

AEZs Africa South of the Sahara (SSA) based on the methodology developed by FAO and IIASA, using the scheme with 5 classes

Mean.Annual.Precip

mean annual precipitation (mm) derived from the CHIRPS 2.0 dataset API at 0.05&deg resolution

Mean.Annual.Temp

mean annual temperature (C) derived from the NASA POWER dataset API at 0.5&deg resolution

CLY

weight percentage of clay particles (<0.0002 mm)

SLT

weight percentage of silt particles (0.0002-0.05 mm)

SND

weight percentage of the sand particles (0.05-2 mm)

...

ERA.Compiled

Format

An object of class data.table (inherits from data.frame) with 112859 rows and 135 columns.

Source

ERA Project Team (2021)