Outcome Codes (Version 23.01.2023)

Code

unique alpha numeric code

Pillar

concepts in the highest level in the outcome hierarchy

Pillar.Code

currently a duplicate of Pillar

Subpillar

concepts in the intermediate level of the outcome hierarchy. Subpillar is nested under Pillar

Subpillar.Code

short code corresponding to the Subpillar name for use in plotting

Indicator

concepts in the intermediate level of the outcome hierarchy. Indicator is nested under Subpillar

Indicator.Code

short code corresponding to the Indicator name for use in plotting

Subindicator

concepts in the lowest and most detailed level of the outcome hierarchy. Subindicator is nested under Indicator

Subindicator.Short

shorter name for the Subindicator

Subindicator.Code

short code corresponding to the Subindicator name for use in plotting

Definition

description of pthe Subindicator

Notes

free text used to clarify the defition of the Subindicator

Example.units

example of units that are collecting for that Subindicator

Original.Outcome

currently NA

Negative.Values

indicates whether the outcome can have negative values; N- No and Y-Yes

Sign

indicates whether the outcome is positive or negative

TC.Ratio

this is the ratio between the outcome for the treatment and control. TC = TreatmeantControl

...

OutcomeCodes

Format

An object of class data.table (inherits from data.frame) with 115 rows and 19 columns.

Source

ERA Project Team (2021)